+++ tuxcap/lib/DescParser.cpp
@@ -48,7 +48,7 @@ std::string DescParser::Unquote(const std::string& theQuotedString)
 		std::string aLiteralString;
 		bool lastWasQuote = false;
 					
-		for (ulong i = 0; i < theQuotedString.length(); i++)
+		for (uint32_t i = 0; i < theQuotedString.length(); i++)
 		{
 			if (theQuotedString[i] == aQuoteChar)
 			{
@@ -74,7 +74,7 @@ bool DescParser::GetValues(ListDataElement* theSource, ListDataElement* theValue
 {
 	theValues->mElementVector.clear();
 	
-	for (ulong aSourceNum = 0; aSourceNum < theSource->mElementVector.size(); aSourceNum++)
+	for (uint32_t aSourceNum = 0; aSourceNum < theSource->mElementVector.size(); aSourceNum++)
 	{
 		if (theSource->mElementVector[aSourceNum]->mIsList)
 		{
@@ -130,7 +130,7 @@ std::string DescParser::DataElementToString(DataElement* theDataElement)
 
 		std::string aString = "(";
 
-		for (ulong i = 0; i < aListDataElement->mElementVector.size(); i++)
+		for (uint32_t i = 0; i < aListDataElement->mElementVector.size(); i++)
 		{
 			if (i != 0)
 				aString += ", ";
@@ -219,7 +219,7 @@ bool DescParser::DataToStringVector(DataElement* theSource, StringVector* theStr
 		aValues = (ListDataElement*) aDataElement;
 	}	
 
-	for (ulong i = 0; i < aValues->mElementVector.size(); i++)
+	for (uint32_t i = 0; i < aValues->mElementVector.size(); i++)
 	{
 		if (aValues->mElementVector[i]->mIsList)
 		{
@@ -262,7 +262,7 @@ bool DescParser::DataToIntVector(DataElement* theSource, IntVector* theIntVector
 	if (!DataToStringVector(theSource, &aStringVector))
 		return false;	
 
-	for (ulong i = 0; i < aStringVector.size(); i++)
+	for (uint32_t i = 0; i < aStringVector.size(); i++)
 	{		
 		int aIntVal;
 		if (!StringToInt(aStringVector[i], &aIntVal))
@@ -282,7 +282,7 @@ bool DescParser::DataToDoubleVector(DataElement* theSource, DoubleVector* theDou
 	if (!DataToStringVector(theSource, &aStringVector))
 		return false;	
 
-	for (ulong i = 0; i < aStringVector.size(); i++)
+	for (uint32_t i = 0; i < aStringVector.size(); i++)
 	{		
 		double aDoubleVal;
 		if (!StringToDouble(aStringVector[i], &aDoubleVal))
