+++ tuxcap/lib/WidgetManager.cpp
@@ -180,7 +180,7 @@ void WidgetManager::FlushDeferredOverlayWidgets(int theMaxPriority)
 	}
 }
 
-void WidgetManager::DoMouseUps(Widget* theWidget, ulong theDownCode)
+void WidgetManager::DoMouseUps(Widget* theWidget, uint32_t theDownCode)
 {
 	int aClickCountTable[3] = { 1,-1, 3 };
 	for (int i = 0; i < 3; i++)
